requirements.txt
python_ipmi==0.5.2
paho_mqtt==1.6.1
PyYAML==6.0
